+++ app/src/androidTest/java/io/github/mgdx/rouelibre/data/addresses/AddressIndexTest.kt
@@ -1,6 +1,5 @@
 package io.github.mgdx.rouelibre.data.addresses
 
-import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry